Moved structs/types to spv_state file
